requirements.txt
torch
transformers
Pillow
flask
